+++ contracts/hackatom/tests/integration.rs
@@ -53,7 +53,7 @@ fn make_init_msg() -> (InstantiateMsg, String) {
 #[test]
 fn proper_initialization() {
     let mut deps = mock_instance(WASM, &[]);
-    assert_eq!(deps.required_features.len(), 0);
+    assert_eq!(deps.required_features().len(), 0);
 
     let verifier = String::from("verifies");
     let beneficiary = String::from("benefits");