+++ contracts/staking/tests/integration.rs
@@ -86,8 +86,8 @@ fn proper_initialization() {
     );
     let (instance_options, memory_limit) = mock_instance_options();
     let mut deps = Instance::from_code(WASM, backend, instance_options, memory_limit).unwrap();
-    assert_eq!(deps.required_features.len(), 1);
-    assert!(deps.required_features.contains("staking"));
+    assert_eq!(deps.required_features().len(), 1);
+    assert!(deps.required_features().contains("staking"));
 
     let creator = String::from("creator");
     let msg = InstantiateMsg {