+++ CHANGELOG.md
@@ -66,6 +66,8 @@ and this project adheres to
   downstream code).
 - cosmwasm-std: `Event::attr` has been renamed to `Event::add_attribute` for
   consistency with other types like `Response`.
+- cosmwasm-vm: `Instance::required_features` changed from a property to a getter
+  method.
 
 [#995]: https://github.com/CosmWasm/cosmwasm/pull/995